requirements.txt
streamlit
supabase
google-api-python-client
google-auth
google-auth-httplib2
rapidfuzz
pandas
